correct propagation of changes
The previous implementation was incorrect and didn't make sure that
"changes" are propagated to all the influenced statements.

The essential issue here is that, without usage information of argument
or SSA values, this backward analysis doesn't fully understand which
statement can be affected by this statement (by contrast a forward
analysis can just use `Goto` and `GotoIfNot` in order to respect the
control flow, and they are always available at inference).

Let me explain the problem with an example; a current implementation end
up with a single traversal and will produce the following result
```
%1 = alloc
%2 = escape!(%1)        # %1 => escape
%3 = alloc
%4 = setfield!(%1, %3)  # (we have to come back to this statement after %2)
     return %3          # %3 => return escape
```

This is because the change of `%2` doesn't propagate to `%4`.
To fix this "correctly", we need to construct a mapping of
(argument, SSA values) -> (argument usages, SSA values usages),
but this PR right now simply keeps linear traversing until every
statement gets converged and it should be a valid fix (but maybe
less efficient).
We may want to implement the "correct" fix and compare their performances.